+++ llvm/docs/ReleaseNotes.rst
@@ -117,6 +117,7 @@ Changes to the ARM Backend
 --------------------------
 
 * Added support for Cortex-M52 CPUs.
+* Added execute-only support for Armv6-M.
 
 Changes to the AVR Backend
 --------------------------
@@ -153,6 +154,7 @@ Changes to the RISC-V Backend
   "SiFive Custom Instruction Extension" as SiFive defines it. The LLVM project
   needs to work with SiFive to define and document real extension names for
   individual CSRs and instructions.
+* ``-mcpu=sifive-p450`` was added.
 
 Changes to the WebAssembly Backend
 ----------------------------------